// src/app/models/employment.model.ts
export class Employment {

    role: String;
    adscription: String;
    isObliged: Boolean;
    contractType: String;
    posessionDate: Date;
    nominaKey: String;
    office: String;
    position: String;
    ext: Number;
    status: Boolean;
    permitIniDate: Date;
    permitEndDate: Date;
    banDate: Date;
    createdAt: Date;
    modifiedAt: Date;
    modifierId: String;
    creatorId: String;

}